Conflicts resloved:
	chromium/base/BUILD.gn
	chromium/base/allocator/partition_allocator/page_allocator.cc
	chromium/base/export_template.h
	chromium/base/files/file_posix.cc
	chromium/base/files/file_util_posix.cc
	chromium/base/files/memory_mapped_file_posix.cc
	chromium/base/i18n/icu_util.cc
	chromium/base/i18n/string_compare.cc
	chromium/base/memory/shared_memory.h
	chromium/base/memory/shared_memory_handle.h
	chromium/base/memory/shared_memory_tracker.cc
	chromium/base/message_loop/message_loop.h
	chromium/base/metrics/field_trial.cc
	chromium/base/metrics/field_trial.h
	chromium/base/process/launch_posix.cc
	chromium/base/process/memory.cc
	chromium/base/profiler/test_support_library.cc
	chromium/base/time/time_exploded_posix.cc
	chromium/build/build_config.h
	chromium/build/config/compiler/BUILD.gn
	chromium/build/config/linux/pkg-config.py
	chromium/build/config/sanitizers/BUILD.gn
	chromium/build/toolchain/gcc_toolchain.gni
	chromium/components/autofill/core/common/autofill_regexes.cc
	chromium/components/crash/core/common/BUILD.gn
	chromium/components/spellcheck/renderer/spellcheck_worditerator.cc
	chromium/components/url_formatter/idn_spoof_checker.cc
	chromium/components/variations/proto/study.proto
	chromium/content/browser/browser_process_sub_thread.cc
	chromium/content/browser/child_process_launcher_helper_posix.cc
	chromium/content/shell/BUILD.gn
	chromium/crypto/nss_util.cc
	chromium/device/gamepad/public/cpp/gamepad.h
	chromium/extensions/browser/api/messaging/message_service.cc
	chromium/gin/v8_platform.cc
	chromium/ipc/ipc_message_utils.cc
	chromium/ipc/native_handle_type_converters.cc
	chromium/media/cdm/cdm_manager_export.h
	chromium/mojo/core/channel.cc
	chromium/mojo/core/channel.h
	chromium/mojo/core/channel_posix.cc
	chromium/mojo/core/node_controller.cc
	chromium/mojo/core/platform_handle_in_transit.cc
	chromium/mojo/core/platform_handle_in_transit.h
	chromium/mojo/public/c/system/platform_handle.h
	chromium/mojo/public/cpp/platform/platform_handle.cc
	chromium/mojo/public/cpp/platform/platform_handle.h
	chromium/mojo/public/cpp/system/platform_handle.cc
	chromium/mojo/public/interfaces/bindings/native_struct.mojom
	chromium/net/BUILD.gn
	chromium/net/base/network_interfaces_posix.cc
	chromium/net/cert/cert_verify_proc.cc
	chromium/net/cert/cert_verify_proc_nss.cc
	chromium/net/cert/known_roots_nss.cc
	chromium/net/dns/host_resolver_impl.cc
	chromium/net/dns/host_resolver_proc.cc
	chromium/net/socket/socks5_client_socket.cc
	chromium/net/third_party/quic/platform/impl/quic_ip_address_impl.cc
	chromium/printing/BUILD.gn
	chromium/printing/backend/print_backend.h
	chromium/printing/backend/test_print_backend.cc
	chromium/printing/backend/test_print_backend.h
	chromium/printing/buildflags/buildflags.gni
	chromium/services/device/geolocation/location_arbitrator.cc
	chromium/services/device/time_zone_monitor/BUILD.gn
	chromium/services/device/time_zone_monitor/time_zone_monitor.cc
	chromium/services/device/wake_lock/power_save_blocker/BUILD.gn
	chromium/services/tracing/public/mojom/tracing.mojom
	chromium/third_party/abseil-cpp/absl/base/internal/spinlock_posix.inc
	chromium/third_party/blink/renderer/platform/fonts/font_cache.h
	chromium/third_party/blink/renderer/platform/fonts/skia/font_cache_skia.cc
	chromium/third_party/blink/renderer/platform/heap/asm/SaveRegisters_x86.asm
	chromium/third_party/blink/renderer/platform/wtf/stack_util.cc
	chromium/third_party/blink/renderer/platform/wtf/text/wtf_string.h
	chromium/third_party/boringssl/src/crypto/fipsmodule/rand/urandom.c
	chromium/third_party/dawn/third_party/glad/include/glad/glad.h
	chromium/third_party/libaom/source/libaom/av1/encoder/x86/temporal_filter_apply_sse2.asm
	chromium/third_party/pdfium/core/fxcrt/cfx_datetime.cpp
	chromium/third_party/pdfium/core/fxcrt/cfx_fileaccess_posix.h
	chromium/third_party/pdfium/core/fxcrt/fx_system.h
	chromium/third_party/pdfium/core/fxge/fx_ge_linux.cpp
	chromium/third_party/pdfium/fxjs/cjs_publicmethods.cpp
	chromium/third_party/protobuf/proto_library.gni
	chromium/third_party/skia/BUILD.gn
	chromium/third_party/skia/gn/toolchain/BUILD.gn
	chromium/third_party/skia/include/core/SkPreConfig.h
	chromium/third_party/sqlite/amalgamation/rename_exports.h
	chromium/third_party/sqlite/src/amalgamation_dev/shell/shell.c
	chromium/third_party/sqlite/src/amalgamation_dev/sqlite3.c
	chromium/third_party/sqlite/src/amalgamation_dev/sqlite3.h
	chromium/third_party/sqlite/src/configure
	chromium/third_party/sqlite/src/src/os_unix.c
	chromium/third_party/usrsctp/usrsctplib/usrsctplib/netinet/sctp_usrreq.c
	chromium/third_party/webrtc/rtc_base/ifaddrs_converter.cc
	chromium/third_party/webrtc/rtc_base/ip_address.cc
	chromium/third_party/webrtc/rtc_base/network.cc
	chromium/third_party/webrtc/rtc_base/platform_file.cc
	chromium/third_party/yasm/yasm_assemble.gni
	chromium/tools/ipc_fuzzer/message_dump/message_dump.cc
	chromium/tools/variations/fieldtrial_to_struct.py
	chromium/ui/base/BUILD.gn
	chromium/ui/base/ime/linux/ui_base_ime_linux_export.h
	chromium/ui/base/ime/ui_base_ime_export.h
	chromium/ui/base/ime/ui_base_ime_types_export.h
	chromium/ui/base/x/ui_base_x_export.h
	chromium/ui/events/ozone/evdev/events_ozone_evdev_export.h
	chromium/ui/events/ozone/events_ozone_export.h
	chromium/ui/events/ozone/layout/events_ozone_layout_export.h
	chromium/ui/gfx/BUILD.gn
	chromium/ui/gfx/font_render_params.h
	chromium/ui/gl/gl_fence.cc
	chromium/ui/keyboard/keyboard_export.h
	chromium/ui/ozone/ozone_base_export.h
	chromium/ui/ozone/ozone_export.h
	chromium/ui/platform_window/android/android_window_export.h
	chromium/ui/platform_window/mojo/mojo_ime_export.h
	chromium/ui/views/mus/mus_export.h
	chromium/ui/views_bridge_mac/views_bridge_mac_export.h
	chromium/v8/BUILD.gn
	chromium/v8/include/v8config.h
	chromium/v8/src/base/platform/platform-posix.cc
	chromium/v8/src/objects/js-list-format.cc
	chromium/v8/src/objects/js-list-format.h
	chromium/v8/src/objects/js-number-format.cc
	chromium/v8/src/snapshot/embedded-file-writer.cc
	gn/build/gen.py
	gn/tools/gn/args.cc
	gn/tools/gn/filesystem_utils.cc
	gn/tools/gn/setup.cc
	gn/tools/gn/substitution_writer_unittest.cc
	ninja/configure.py
	ninja/src/disk_interface.cc
	ninja/src/subprocess-posix.cc

+++ chromium/components/history/core/browser/expire_history_backend.cc
@@ -19,6 +19,7 @@
 #include "base/logging.h"
 #include "base/metrics/histogram_macros.h"
 #include "base/sequenced_task_runner.h"
+#include "base/stl_util.h"
 #include "build/build_config.h"
 #include "components/history/core/browser/history_backend_client.h"
 #include "components/history/core/browser/history_backend_notifier.h"
@@ -169,8 +170,7 @@ ExpireHistoryBackend::ExpireHistoryBackend(
       main_db_(nullptr),
       thumb_db_(nullptr),
       backend_client_(backend_client),
-      task_runner_(task_runner),
-      weak_factory_(this) {
+      task_runner_(task_runner) {
   DCHECK(notifier_);
 }
 
@@ -183,11 +183,12 @@ void ExpireHistoryBackend::SetDatabases(HistoryDatabase* main_db,
   thumb_db_ = thumb_db;
 }
 
-void ExpireHistoryBackend::DeleteURL(const GURL& url) {
-  DeleteURLs(std::vector<GURL>(1, url));
+void ExpireHistoryBackend::DeleteURL(const GURL& url, base::Time end_time) {
+  DeleteURLs({url}, end_time);
 }
 
-void ExpireHistoryBackend::DeleteURLs(const std::vector<GURL>& urls) {
+void ExpireHistoryBackend::DeleteURLs(const std::vector<GURL>& urls,
+                                      base::Time end_time) {
   if (!main_db_)
     return;
 
@@ -206,16 +207,24 @@ void ExpireHistoryBackend::DeleteURLs(const std::vector<GURL>& urls) {
     // Collect all the visits and delete them. Note that we don't give up if
     // there are no visits, since the URL could still have an entry that we
     // should delete.
-    VisitVector visits;
-    main_db_->GetVisitsForURL(url_row.id(), &visits);
+    VisitVector visits_to_delete;
+    main_db_->GetVisitsForURL(url_row.id(), &visits_to_delete);
+    size_t total_visits = visits_to_delete.size();
+    if (!end_time.is_null() && !end_time.is_max()) {
+      // Remove all items that should not be deleted from |visits_to_delete|.
+      base::EraseIf(visits_to_delete,
+                    [=](auto& v) { return v.visit_time > end_time; });
+    }
+    DeleteVisitRelatedInfo(visits_to_delete, &effects);
 
-    DeleteVisitRelatedInfo(visits, &effects);
-
-    // We skip ExpireURLsForVisits (since we are deleting from the
-    // URL, and not starting with visits in a given time range). We
-    // therefore need to call the deletion and favicon update
-    // functions manually.
-    DeleteOneURL(url_row, is_pinned, &effects);
+    // Remove the URL if all visits have been removed.
+    if (visits_to_delete.size() == total_visits) {
+      // We skip ExpireURLsForVisits (since we are deleting from the
+      // URL, and not starting with visits in a given time range). We
+      // therefore need to call the deletion and favicon update
+      // functions manually.
+      DeleteOneURL(url_row, is_pinned, &effects);
+    }
   }
 
   DeleteFaviconsIfPossible(&effects);
@@ -323,6 +332,43 @@ void ExpireHistoryBackend::ExpireHistoryBeforeForTesting(base::Time end_time) {
   ParanoidExpireHistory();
 }
 
+void ExpireHistoryBackend::ClearOldOnDemandFaviconsIfPossible(
+    base::Time expiration_threshold) {
+  if (!thumb_db_)
+    return;
+
+  // Extra precaution to avoid repeated calls to GetOldOnDemandFavicons() close
+  // in time, since it can be fairly expensive.
+  if (expiration_threshold <
+      last_on_demand_expiration_threshold_ +
+          base::TimeDelta::FromHours(kClearOnDemandFaviconsIntervalHours)) {
+    return;
+  }
+
+  last_on_demand_expiration_threshold_ = expiration_threshold;
+
+  std::map<favicon_base::FaviconID, IconMappingsForExpiry> icon_mappings =
+      thumb_db_->GetOldOnDemandFavicons(expiration_threshold);
+  DeleteEffects effects;
+
+  for (auto id_and_mappings_pair : icon_mappings) {
+    favicon_base::FaviconID icon_id = id_and_mappings_pair.first;
+    const IconMappingsForExpiry& mappings = id_and_mappings_pair.second;
+
+    if (backend_client_ &&
+        IsAnyURLPinned(backend_client_, mappings.page_urls)) {
+      continue;
+    }
+
+    thumb_db_->DeleteFavicon(icon_id);
+    thumb_db_->DeleteIconMappingsForFaviconId(icon_id);
+    effects.deleted_favicons.insert(mappings.icon_url);
+  }
+
+  BroadcastNotifications(&effects, DELETION_EXPIRED,
+                         DeletionTimeRange::Invalid(), base::nullopt);
+}
+
 void ExpireHistoryBackend::InitWorkQueue() {
   DCHECK(work_queue_.empty()) << "queue has to be empty prior to init";
 
@@ -401,7 +447,7 @@ VisitVector ExpireHistoryBackend::GetVisitsAndRedirectParents(
     const VisitVector& visits) {
   base::flat_set<VisitID> seen_visits;
   VisitVector visits_and_redirects;
-  for (const auto v : visits) {
+  for (const auto& v : visits) {
     VisitRow current_visit = v;
     do {
       if (!seen_visits.insert(current_visit.visit_id).second)
@@ -576,43 +622,6 @@ void ExpireHistoryBackend::DoExpireIteration() {
   }
 
   ScheduleExpire();
-}
-
-void ExpireHistoryBackend::ClearOldOnDemandFaviconsIfPossible(
-    base::Time expiration_threshold) {
-  if (!thumb_db_)
-    return;
-
-  // Extra precaution to avoid repeated calls to GetOldOnDemandFavicons() close
-  // in time, since it can be fairly expensive.
-  if (expiration_threshold <
-      last_on_demand_expiration_threshold_ +
-          base::TimeDelta::FromHours(kClearOnDemandFaviconsIntervalHours)) {
-    return;
-  }
-
-  last_on_demand_expiration_threshold_ = expiration_threshold;
-
-  std::map<favicon_base::FaviconID, IconMappingsForExpiry> icon_mappings =
-      thumb_db_->GetOldOnDemandFavicons(expiration_threshold);
-  DeleteEffects effects;
-
-  for (auto id_and_mappings_pair : icon_mappings) {
-    favicon_base::FaviconID icon_id = id_and_mappings_pair.first;
-    const IconMappingsForExpiry& mappings = id_and_mappings_pair.second;
-
-    if (backend_client_ &&
-        IsAnyURLPinned(backend_client_, mappings.page_urls)) {
-      continue;
-    }
-
-    thumb_db_->DeleteFavicon(icon_id);
-    thumb_db_->DeleteIconMappingsForFaviconId(icon_id);
-    effects.deleted_favicons.insert(mappings.icon_url);
-  }
-
-  BroadcastNotifications(&effects, DELETION_EXPIRED,
-                         DeletionTimeRange::Invalid(), base::nullopt);
 }
 
 bool ExpireHistoryBackend::ExpireSomeOldHistory(
